refactor(render-utils): migrate operation record logic to recordUtils from renderUtils
