+++ src/worker/utils/record.ts
@@ -1,18 +1,28 @@
 import type {
 	BroomRecord,
 	BucketRecord,
+	CanvasType,
 	CircleRecord,
 	EraserRecord,
 	LineRecord,
 	OpRecord,
 	PencilRecord,
 	RecordMessagePayload,
+	RedoOrUndoMessagePayload,
 	SquareRecord,
 } from "@/types";
 import { ToolTypeEnum } from "@/types";
-import { useRecords } from "../signals";
+import { useRecords, useRender } from "../signals";
+import * as renderUtils from "./render";
+
+interface ReplayRecordsConfig {
+	tabId: string;
+	frameId: string;
+	canvasType?: CanvasType;
+}
 
 const {
+	getColor,
 	getRecords,
 	getColorIndex,
 	getFrameIndex,
@@ -22,7 +32,13 @@ const {
 	addRecordToUndoStack,
 	clearRedoStack,
 	clearRecordPoints,
+	getUndoStack,
+	popRedoStack,
+	popUndoStack,
+	addRecordToRedoStack,
 } = useRecords();
+
+const { resetColorPositionMap, clearVisited } = useRender();
 
 const makePencilRecord = (
 	payload: RecordMessagePayload,
@@ -211,4 +227,196 @@ export const record = (payload: RecordMessagePayload) => {
 	clearRedoStack(tabId);
 	addRecordToUndoStack(tabId, record);
 	return true;
+};
+
+export const getRecordsWithFrameId = (tabId: string, frameId: string) => {
+	const undoStack = getUndoStack(tabId);
+	return undoStack.filter((record) => {
+		const [toolType] = record;
+		const frameIndex = toolType === ToolTypeEnum.Eraser ? record[1] : record[2];
+		const _frameIndex = getFrameIndex(tabId, frameId);
+		return _frameIndex === frameIndex;
+	});
+};
+
+export const redo = (payload: RedoOrUndoMessagePayload) => {
+	const { tabId, frameId } = payload;
+	const record = popRedoStack(tabId);
+	if (!record) return;
+
+	addRecordToUndoStack(tabId, record);
+	resetColorPositionMap();
+	clearVisited();
+	replayRecords(getUndoStack(tabId), {
+		tabId,
+		frameId,
+	});
+};
+
+export const undo = (payload: RedoOrUndoMessagePayload) => {
+	const { tabId, frameId } = payload;
+	const record = popUndoStack(tabId);
+
+	if (!record) return;
+
+	addRecordToRedoStack(tabId, record);
+	resetColorPositionMap();
+	clearVisited();
+	replayRecords(getUndoStack(tabId), {
+		tabId,
+		frameId,
+	});
+};
+
+export const replayRecords = (
+	records: OpRecord[],
+	config: ReplayRecordsConfig,
+) => {
+	renderUtils.clearAllPixels({ canvasType: "main" });
+
+	for (const record of records) {
+		const [toolType] = record;
+
+		switch (toolType) {
+			case ToolTypeEnum.Pencil:
+				replayPencilRecord(record as PencilRecord, config);
+				break;
+			case ToolTypeEnum.Eraser:
+				replayEraserRecord(record as EraserRecord, config);
+				break;
+			case ToolTypeEnum.Line:
+				replayLineRecord(record as LineRecord, config);
+				break;
+			case ToolTypeEnum.Square:
+				replaySquareRecord(record as SquareRecord, config);
+				break;
+			case ToolTypeEnum.Circle:
+				replayCircleRecord(record as CircleRecord, config);
+				break;
+			case ToolTypeEnum.Ellipse:
+				replayCircleRecord(record as CircleRecord, config);
+				break;
+			case ToolTypeEnum.Bucket:
+				replayBucketRecord(record as BucketRecord, config);
+				break;
+			case ToolTypeEnum.Broom:
+				replayClearAllPixelsRecord();
+				break;
+		}
+
+		clearVisited();
+	}
+};
+
+const replayPencilRecord = (
+	record: PencilRecord,
+	config: ReplayRecordsConfig,
+) => {
+	const [_, colorIndex, __, pixelSize, points] = record;
+	const { tabId, canvasType = "main" } = config;
+	for (const [x, y, drawCounts] of points) {
+		for (let i = 0; i < drawCounts; i++) {
+			renderUtils.fillRect({
+				toolType: ToolTypeEnum.Pencil,
+				position: { x, y },
+				canvasType,
+				pixelSize,
+				pixelColor: getColor(tabId, colorIndex),
+				isReplay: true,
+			});
+		}
+	}
+};
+
+const replayEraserRecord = (
+	record: EraserRecord,
+	config: ReplayRecordsConfig,
+) => {
+	const [_, __, pixelSize, points] = record;
+	const { canvasType = "main" } = config;
+	for (const [x, y] of points) {
+		renderUtils.clearRect({
+			toolType: ToolTypeEnum.Eraser,
+			position: { x, y },
+			canvasType,
+			pixelSize,
+			isReplay: true,
+		});
+	}
+};
+
+const replayLineRecord = (record: LineRecord, config: ReplayRecordsConfig) => {
+	const [_, colorIndex, __, pixelSize, points] = record;
+	const { tabId, canvasType = "main" } = config;
+	const [startPoint, endPoint] = points;
+	const [startX, startY] = startPoint;
+	const [endX, endY] = endPoint;
+
+	renderUtils.drawBresenhamLine({
+		toolType: ToolTypeEnum.Line,
+		canvasType,
+		lineStartPosition: { x: startX, y: startY },
+		lineEndPosition: { x: endX, y: endY },
+		pixelColor: getColor(tabId, colorIndex),
+		pixelSize,
+	});
+};
+
+const replaySquareRecord = (
+	record: SquareRecord,
+	config: ReplayRecordsConfig,
+) => {
+	const [_, colorIndex, __, pixelSize, points] = record;
+	const { tabId, canvasType = "main" } = config;
+	const [startPoint, endPoint] = points;
+	const [startX, startY] = startPoint;
+	const [endX, endY] = endPoint;
+
+	renderUtils.drawSquare({
+		canvasType,
+		squareStartPosition: { x: startX, y: startY },
+		squareEndPosition: { x: endX, y: endY },
+		pixelSize,
+		pixelColor: getColor(tabId, colorIndex),
+	});
+};
+
+const replayCircleRecord = (
+	record: CircleRecord,
+	config: ReplayRecordsConfig,
+) => {
+	const [toolType, colorIndex, __, pixelSize, points] = record;
+	const { tabId, canvasType = "main" } = config;
+	const [startPoint, endPoint] = points;
+	const [startX, startY] = startPoint;
+	const [endX, endY] = endPoint;
+
+	renderUtils.drawCircle({
+		toolType,
+		canvasType,
+		circleStartPosition: { x: startX, y: startY },
+		circleEndPosition: { x: endX, y: endY },
+		pixelSize,
+		pixelColor: getColor(tabId, colorIndex),
+	});
+};
+
+const replayBucketRecord = (
+	record: BucketRecord,
+	config: ReplayRecordsConfig,
+) => {
+	const [_, colorIndex, __, pixelSize, point] = record;
+	const { tabId, canvasType = "main" } = config;
+	const [x, y] = point;
+
+	renderUtils.fillBucket({
+		canvasType,
+		replacementColor: getColor(tabId, colorIndex),
+		pixelSize,
+		position: { x, y },
+	});
+};
+
+const replayClearAllPixelsRecord = () => {
+	renderUtils.clearAllPixels({ canvasType: "main" });
 };
